Moving ValueOrArray combinators
