+++ packages/graph/hash_graph/lib/graph/src/types/schema/combinator.rs
@@ -1,35 +1,12 @@
 use serde::{Deserialize, Serialize};
 
-use crate::types::{
-    schema::{
-        array::{Array, Itemized},
-        object::ValidateUri,
-        ValidationError,
-    },
-    BaseUri,
-};
+use crate::types::schema::ValidationError;
 
 #[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
 #[serde(untagged, deny_unknown_fields)]
 pub enum Optional<T> {
     None {},
     Some(T),
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(untagged, deny_unknown_fields)]
-pub enum ValueOrArray<T> {
-    Value(T),
-    Array(Itemized<Array, T>),
-}
-
-impl<T: ValidateUri> ValidateUri for ValueOrArray<T> {
-    fn validate_uri(&self, base_uri: &BaseUri) -> error_stack::Result<(), ValidationError> {
-        match self {
-            Self::Value(value) => value.validate_uri(base_uri),
-            Self::Array(array) => array.items().validate_uri(base_uri),
-        }
-    }
 }
 
 #[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
@@ -105,26 +82,6 @@ mod tests {
         #[test]
         fn some() -> Result<(), serde_json::Error> {
             check(&Optional::Some("value".to_owned()), json!("value"))
-        }
-    }
-
-    mod value_or_array {
-        use super::*;
-
-        #[test]
-        fn value() -> Result<(), serde_json::Error> {
-            check(&ValueOrArray::Value("value".to_owned()), json!("value"))
-        }
-
-        #[test]
-        fn array() -> Result<(), serde_json::Error> {
-            check(
-                &ValueOrArray::Array(Itemized::new(Array::default(), "string".to_owned())),
-                json!({
-                    "type": "array",
-                    "items": "string",
-                }),
-            )
         }
     }
